Decouple photo questions from filesystem.
