+++ lib/tenant-interview.ts
@@ -14,7 +14,6 @@ import {
   NonBlankQuestion,
   YesNoQuestion,
   DateQuestion,
-  PhotoQuestion,
 } from './question';
 
 const RENTAL_HISTORY_FOLLOWUP_DAYS = 7;
@@ -73,7 +72,7 @@ export class TenantInterview extends Interview<Tenant> {
       const details = await this.io.askMany({
         dateReceived: new DateQuestion('When did you receive your rental history?'),
         isRentStabilized: new YesNoQuestion('Are you rent stabilized?'),
-        photo: new PhotoQuestion('Please submit a photograph of your rental history.')
+        photo: this.io.createPhotoQuestion('Please submit a photograph of your rental history.')
       });
       return {
         status: 'received',
